sunxi: axp20x: Report voltage, current and power for ACIN connector
These measurements for ACIN (adapter input) are just the momentary
values sampled from ADC. But in order to have something more practical,
a lot of values need to be sampled and averaged. Just obtaining a
single output from the 'sensors' tool is not very meaningful,
because the momentary current draw changes very fast and within
a rather wide range (jumping up and down +-50% is not very uncommon).

To make accurate measurements possible at least from the userland,
it is unreasonable to limit the frequency of sysfs hwmon entries
update to just one read per two seconds. So this limitation has
been relaxed to allow polling at 50Hz.

v2: The measurement of ACIN current is done before anything else
    because it is the most volatile parameter. And we want to
    reduce the chances of the results being skewed by the CPU
    activity related to reading the hwmon sysfs entry itself.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>