sunxi: defconfig: Enable ROOT_NFS support
This patch ensures that default configs allow multiple sun4i,
sun5i and sun7i devices boot and work nicely with the same NFS
root. Tested with Cubietruck, Cubieboard2 and A10-Lime.

ROOT_NFS is enabled in sun4i_defconfig. GMAC is enabled in
sun7i_defconfig for Cubietruck support. USB_NET is enabled
for sun5i because it can allow to use USB ethernet dongles.
Also different localversion identifiers are set for in the
config files to avoid a possible clash between the names of
directories with kernel modules.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>