ARM: dts: sunxi: Update simplfb nodes so that u-boot can find them
Review of the u-boot sunxi simplefb patches has led to the decision that
u-boot should not use a specific path to find the nodes as this goes contrary
to how devicetree usually works.

Instead a platform specific compatible + properties should be used for this.

The simplefb bindings have already been updated to reflect this, this patch
brings the sunxi devicetree files in line with the new binding, and the
actual u-boot implementation as it is going upstream.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>
---
Changes in v2:
-Use proper allwinner, vendor prefix instead of sunxi,
-Rename compatible to allwinner,simple-framebuffer to make clear these are
 vendor extensions to the simple-framebuffer bindings

foo