sunxi: defconfig: Use performance cpufreq governor by default
A rather popular problem reported by new users on irc is very
laggy performance out of the box. The default ondemand governor
is too slow to react on increased demand for processing power
without extra tuning. This extra tuning needs extra efforts
from the users and prior knowledge that such tuning might be
necessary.

The most important defconfigs are sun4i_defconfig, sun5i_defconfig
and sun7i_defconfig because they are mentioned in
    http://linux-sunxi.org/Linux_Kernel#Compilation

This patch changes defconfigs to have the following set of governors:
  1. 'performance' (default)
  2. 'userspace' (for benchmarking purposes)
  3. 'interactive' (a replacement for 'ondemand' with more sane
     default behavior)

One more reason not to use the 'ondemand' governor by default is
that the power saving provided by it is not particularly good:
    https://www.mail-archive.com/linux-sunxi@googlegroups.com/msg00492.html
    https://www.mail-archive.com/linux-sunxi@googlegroups.com/msg00678.html
    http://www.cubieforums.com/index.php/topic,1413.msg8745.html#msg8745

It basically wastes a lot of performance for almost nothing.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>