sunxi: mali: Handle both framebuffers in the Mali validation code
The Mali kernel driver allows the userland code to map a buffer
in physical memory (via the MALI_IOC_MEM_MAP_EXT ioctl). This is
normally used for mapping the framebuffer in the userland Mali
blob and directly accessing it for rendering graphics. Because
of security reasons, the kernel driver has a validation check,
which only permits one chunk of physical memory to be mapped
this way. And the Mali driver initialization code configures the
address and the size of this chunk.

Now the problem is that the sunxi hardware has not just one, but
up to two framebuffers (for example, one VGA and one HDMI monitor).
Previously both framebuffers were allocated from the disp reservation
area. So the Mali validation code was just allowing access to the
whole disp reservation area. But with CMA support, the framebuffers
may be allocated anywhere in the CMA area. And now we have two
disconnected chunks of physical memory, which should be both
supported by the MALI_IOC_MEM_MAP_EXT ioctl.

This patch modifies the MALI_IOC_MEM_MAP_EXT validation code to
support both of the sunxi disp framebuffers.

PS. If the kernel driver does not allow to map physical memory via
the MALI_IOC_MEM_MAP_EXT ioctl at all, then the Mali userland blob
still can work and is able to render 3D graphics. However this
is done by using memcpy to copy data into the framebuffer for each
frame and the performance becomes ridiculously bad.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>