sunxi: cedar: Allocate memory buffer only when /dev/cedar_dev is open
Previously the 80MB memory buffer was allocated from CMA when the
cedar module is loaded and kept in use until the module is removed.
Now the buffer is only allocated when /dev/cedar_dev is opened and
deallocated when it is closed.

Additionally, the kernel cmdline parameter 'sunxi_ve_mem_reserve'
can be used to change the cedar memory buffer size to something
larger than 80MB. Since libvdpau-sunxi implemented subtitles
support, it needs more memory and 80MB may be not enough.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>