sunxi: defconfig: Build the Mali driver statically
Sometimes people are just forgetting to copy their kernel modules
to the device. Building the Mali kernel driver statically ensures
that people can't make this mistake anymore and removes one of the
potential obstacles on the way to getting 3D accelatation up and
running.

It is also convenient to be able to include the lima-memtester
test program into an initrd image without bothering about adding
kernel modules to the initrd image too.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>