ARM: sun8i: Add SMP support for the Allwinner H3
The H3 is a quad Cortex-A7. Add the logic to use the IPs used to
control the CPU configuration and the CPU power so that we can
bring up secondary CPUs at boot.

CPUCFG in H3 is similar to CPUCFG in A31, but not exactly the
same. H3 PRCM is not documented at all. The SMP code in the
Allwinner SDK is guarded by the same ifdefs for A31 and H3, so
at least bringing up secondary CPU cores is likely compatible
between these two SoC variants.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>